consensus/exit_probability_randomizer - fix tests due to int64 missuse

+++ consensus/cw_rating_calculator/cw_rating_dfs_impl.c
@@ -15,12 +15,12 @@
 
 static retcode_t cw_rating_dfs_do_dfs_from_db(
     const cw_rating_calculator_t *const cw_calc, flex_trit_t *entry_point,
-    hash_to_indexed_hash_set_map_t *tx_to_approvers, size_t *subtangle_size,
+    hash_to_indexed_hash_set_map_t *tx_to_approvers, uint64_t *subtangle_size,
     int64_t subtangle_before_timestamp);
 
 static retcode_t cw_rating_dfs_do_dfs_light(
     hash_to_indexed_hash_set_map_t tx_to_approvers, flex_trit_t *ep,
-    bitset_t *visited_bitset, size_t *subtangle_size);
+    bitset_t *visited_bitset, uint64_t *subtangle_size);
 
 void init_cw_calculator_dfs(cw_rating_calculator_base_t *calculator) {
   calculator->vtable = cw_topological_vtable;
@@ -35,9 +35,9 @@ retcode_t cw_rating_calculate_dfs(const cw_rating_calculator_t *const cw_calc,
   out->cw_ratings = NULL;
   hash_to_indexed_hash_set_entry_t *curr_hash_to_approvers_entry = NULL;
   hash_to_indexed_hash_set_entry_t *tmp_hash_to_approvers_entry = NULL;
-  int64_t sub_tangle_size;
-  int64_t max_subtangle_size;
-  size_t bitset_size;
+  uint64_t sub_tangle_size;
+  uint64_t max_subtangle_size;
+  uint64_t bitset_size;
 
   if (!entry_point) {
     return RC_NULL_PARAM;
@@ -100,7 +100,7 @@ retcode_t cw_rating_calculate_dfs(const cw_rating_calculator_t *const cw_calc,
 
 static retcode_t cw_rating_dfs_do_dfs_from_db(
     const cw_rating_calculator_t *const cw_calc, flex_trit_t *entry_point,
-    hash_to_indexed_hash_set_map_t *tx_to_approvers, size_t *subtangle_size,
+    hash_to_indexed_hash_set_map_t *tx_to_approvers, uint64_t *subtangle_size,
     int64_t subtangle_before_timestamp) {
   hash_to_indexed_hash_set_entry_t *curr_tx = NULL;
   size_t curr_approver_index;
@@ -163,7 +163,7 @@ static retcode_t cw_rating_dfs_do_dfs_from_db(
 
 static retcode_t cw_rating_dfs_do_dfs_light(
     hash_to_indexed_hash_set_map_t tx_to_approvers, flex_trit_t *ep,
-    bitset_t *visited_bitset, size_t *subtangle_size) {
+    bitset_t *visited_bitset, uint64_t *subtangle_size) {
   *subtangle_size = 0;
   flex_trit_t *curr_hash = NULL;
   hash_to_indexed_hash_set_entry_t *curr_tx_entry = NULL;
